Revert neovim's LSP
The primary problem I have with it is that neovim's default LSP omnifunc
handler does not support fuzzy matches. The LSP server itself does, but
the omnifunc handler is filtering any matches that doesn't strictly
start with the input prefix.

Callstack:
- https://github.com/neovim/neovim/blob/master/runtime/lua/vim/lsp.lua#L1677
- https://github.com/neovim/neovim/blob/b455e0179b4288c69e6231bfcf8d1c132b78f2fc/runtime/lua/vim/lsp/util.lua#L633
- https://github.com/neovim/neovim/blob/b455e0179b4288c69e6231bfcf8d1c132b78f2fc/runtime/lua/vim/lsp/util.lua#L596-L600

I would like to find a way around this in the future, but for now I have
to revert to vim-go's omnifunc implementation.